Rewrite map data generation using Wikidata-Toolkit
This removes use of ALL old code and rewrites the
generated of the 2 files needed for the wikidata map
in Java using the Wikidata-Toolkit.

Current runs use LARGE ammounts of memory, but if
you have a machine powerfull enough then this will
be no problem.

This also adds the html and js code for the map to
this repo which was not previously stored in git.

All other broken / unused code has also been removed
and will likely slowly be re written.