Just use JSON data files
We initially output the datain a .js file
for use with the online map.

It makes much more sense to output a JSON file
which could easily then be converted to a js file

The JSON file can also then easily be used by other
scripts such as geo2png!