Add 7680x4320 size with rivers.
My idea with this size data file would be to try using it to render the
smaller sizes, simply diviving the coordinates by 2 or 4?

This new large size no includes revier relations as well.
To do this I needed to
 - Pass the width of the canvas into the reiver draw method in order to
 correctly eliminate lines that reach across the canvas.
 - Add the river url to the config

Currently this large size map renders in 13 seconds for me, and the data
is also ordered by density, which means the lighter pixels render first
and the brigers ones fill in toward the end of the render.