Use a service worker for the fetching
In combination with the previous patch drawing from the stream that is
still being received, this allows smooth drawing as the fetching and
processing of data is all done by the service worker.

Currently it appears that the messages for dots and rivers can come a
bit mixed up, which means that the map can be rendered in an odd way as
the same canvas is used.

Time to use multiple canvases I think...