canvases, divide dot and river canvas & draw to correct res
This splits up the river and the main map canvas so that if the river
stuff ends up drawing before the dots, the dots dont then overwrite the
rivers.
This also makes sure that the worker writes to the correct canvas if the
user switches away mid render.

Multiple canvas layers also means that we can start rendering each layer
at the same time (thanks streams and service workers).
This will be grent when we also have even more layers!

This also now means that you can't right click download the map, as the
reiver canvas is in front.
This change also changes the transparancy of the river drawings so they
are less in your face.