More Layers! Generate and render more layers for the big map.
This udpates the notes about data generation, tweaking it slightly to
enable easy generation of data for more property relations.
This also dynamically creates more canvases for more layers of
relations.
Each layer has it's own config now, including color and some other
settings.

The layer URL is now a generic template that the property of the
relation is inserted into. The idea in the future would be that even
this template isn't needed, instead this is got dynamically.