worker, continue to render other canvases when switching around
The onmessage method was changed every time the resolution was switched,
which meant any previous renderings that the user was switching away
from woruld stop and be broken forever.
Now the onmessage method is generic and works for all resolutions.

This could still be improved by having the rendering of a newly selected
resolution start straight away, right now it has to wait for the renders
of previous canvases to finish.
Perhaps multiple workers could work?